feat!: ga the new data tree commands
