+++ messages/tree.export.md
@@ -30,7 +30,7 @@ Directory in which to generate the JSON files; default is current directory.
 
 - Export records retrieved with the specified SOQL query into a single JSON file in the current directory; the command uses your default org:
 
-  <%= config.bin %> <%= command.id %> --query "SELECT Id, Name, (SELECT Name, Address__c FROM Properties__r) FROM Broker__c"
+  <%= config.bin %> <%= command.id %> --query "SELECT Id, Name, (SELECT Name, Address**c FROM Properties**r) FROM Broker\_\_c"
 
 - Export data using a SOQL query in the "query.txt" file and generate JSON files for each object and a plan that aggregates them:
 
@@ -43,3 +43,11 @@ Directory in which to generate the JSON files; default is current directory.
 # PrefixSlashError
 
 `--prefix` cannot contain a forward slash or backslash.
+
+# PlanJsonWarning
+
+After Nov 10, 2024, the JSON output for `--plan` will no longer include the `saveRefs` and `resolveRefs` properties.
+
+# LegacyDeprecation
+
+After Nov 10, 2024, this command will no longer be available. Use `data export tree`.
